Ai resuma/requirements.txt
Flask==3.0.0
scikit-learn==1.3.2
pandas==2.1.4
PyPDF2==3.0.1
python-docx==1.1.0
requests==2.31.0
joblib==1.3.2
werkzeug==3.0.1
matplotlib==3.8.2
seaborn==0.13.0
numpy==1.24.3
